fix(evals): align editor with point-cloud loader
